# siem-lite-openapi.json
{"openapi":"3.1.0","info":{"title":"SIEM-Lite API","description":"Portfolio-grade SIEM backend API","version":"1.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Logs","description":"Log ingestion and parsing"},{"name":"Audit Logs","description":"Security audit trail"},{"name":"Indicators of Compromise","description":"Threat intelligence IOC management"},{"name":"Alerts","description":"Alert management and analyst workflow"},{"name":"Events","description":"Security events generated from logs"},{"name":"Health","description":"Application health and status"}],"paths":{"/api/iocs/{id}":{"get":{"tags":["Indicators of Compromise"],"summary":"Get IOC by ID","description":"Returns a single Indicator of Compromise by ID","operationId":"getIocById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}},"put":{"tags":["Indicators of Compromise"],"summary":"Update IOC","description":"Updates an existing Indicator of Compromise","operationId":"updateIoc","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromiseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}},"delete":{"tags":["Indicators of Compromise"],"summary":"Delete IOC","description":"Deletes an Indicator of Compromise","operationId":"deleteIoc","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/iocs/{id}/enable":{"put":{"tags":["Indicators of Compromise"],"summary":"Enable IOC","description":"Enables an IOC for active threat intelligence matching","operationId":"enableIoc","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}}},"/api/iocs/{id}/disable":{"put":{"tags":["Indicators of Compromise"],"summary":"Disable IOC","description":"Disables an IOC from active threat intelligence matching","operationId":"disableIoc","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}}},"/api/incidents/{id}/status":{"put":{"tags":["incident-controller"],"operationId":"updateIncidentStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Incident"}}}}}}},"/api/incidents/{id}/notes":{"put":{"tags":["incident-controller"],"operationId":"updateIncidentNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"notes","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Incident"}}}}}}},"/api/incidents/{id}/assign":{"put":{"tags":["incident-controller"],"operationId":"assignIncident","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"assignedAnalyst","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Incident"}}}}}}},"/api/alerts/{id}/status":{"put":{"tags":["Alerts"],"summary":"Change alert status","operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}}},"/api/alerts/{id}/notes":{"put":{"tags":["Alerts"],"summary":"Add investigation notes to an alert","operationId":"updateNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"note","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}}},"/api/alerts/{id}/assign":{"put":{"tags":["Alerts"],"summary":"Assign analyst to an alert","operationId":"assignAnalyst","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"analyst","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}}},"/api/logs":{"post":{"tags":["Logs"],"summary":"Submit a single log","operationId":"ingestLog","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SecurityEvent"}}}}}}},"/api/logs/bulk":{"post":{"tags":["Logs"],"summary":"Submit multiple logs","operationId":"ingestBulkLogs","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LogRequest"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SecurityEvent"}}}}}}}},"/api/iocs":{"get":{"tags":["Indicators of Compromise"],"summary":"Get all IOCs","description":"Returns all stored Indicators of Compromise","operationId":"getAllIocs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}}},"post":{"tags":["Indicators of Compromise"],"summary":"Create IOC","description":"Creates a new Indicator of Compromise","operationId":"createIoc","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromiseRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IndicatorOfCompromise"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/incidents":{"get":{"tags":["incident-controller"],"operationId":"getAllIncidents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Incident"}}}}}}}},"/api/incidents/{id}":{"get":{"tags":["incident-controller"],"operationId":"getIncidentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Incident"}}}}}}},"/api/health":{"get":{"tags":["Health"],"summary":"Retrieve application health status","operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/events":{"get":{"tags":["Events"],"summary":"Retrieve paginated and filtered security events","operationId":"getEvents","parameters":[{"name":"eventType","in":"query","required":false,"schema":{"type":"string"}},{"name":"sourceIp","in":"query","required":false,"schema":{"type":"string"}},{"name":"username","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageSecurityEvent"}}}}}}},"/api/dashboard":{"get":{"tags":["dashboard-controller"],"summary":"Get dashboard summary metrics","operationId":"getDashboardSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardSummary"}}}}}}},"/api/dashboard/trends":{"get":{"tags":["dashboard-controller"],"summary":"Get dashboard trend analytics","operationId":"getDashboardTrends","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardTrends"}}}}}}},"/api/dashboard/top-usernames":{"get":{"tags":["dashboard-controller"],"summary":"Get top usernames","operationId":"getTopUsernames","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TopUsername"}}}}}}}},"/api/dashboard/top-source-ips":{"get":{"tags":["dashboard-controller"],"summary":"Get top source IP addresses","operationId":"getTopSourceIps","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TopSourceIp"}}}}}}}},"/api/dashboard/top-alert-types":{"get":{"tags":["dashboard-controller"],"summary":"Get top alert types","operationId":"getTopAlertTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TopAlertType"}}}}}}}},"/api/dashboard/timeline":{"get":{"tags":["dashboard-controller"],"summary":"Get dashboard timeline analytics","operationId":"getDashboardTimeline","parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":7}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DashboardTimelinePoint"}}}}}}}},"/api/dashboard/risk-distribution":{"get":{"tags":["dashboard-controller"],"summary":"Get alert risk distribution","operationId":"getRiskDistribution","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RiskDistribution"}}}}}}},"/api/dashboard/mitre-statistics":{"get":{"tags":["dashboard-controller"],"summary":"Get MITRE ATT&CK statistics","operationId":"getMitreStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MitreStatistic"}}}}}}}},"/api/dashboard/ioc-statistics":{"get":{"tags":["dashboard-controller"],"summary":"Get IOC alert statistics","operationId":"getIocStatistics","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IocStatistics"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/audit":{"get":{"tags":["Audit Logs"],"summary":"Retrieve audit logs","operationId":"getAuditLogs","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAuditLog"}}}}}}},"/api/alerts":{"get":{"tags":["Alerts"],"summary":"Retrieve alerts with optional filtering","operationId":"getAlerts","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"assignedTo","in":"query","required":false,"schema":{"type":"string"}},{"name":"mitreTechnique","in":"query","required":false,"schema":{"type":"string"}},{"name":"minRiskScore","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageAlert"}}}}}}},"/api/alerts/{id}/related-events":{"get":{"tags":["Alerts"],"summary":"Retrieve related security events for an alert","operationId":"getRelatedEvents","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SecurityEvent"}}}}}}}}},"components":{"schemas":{"IndicatorOfCompromiseRequest":{"type":"object","properties":{"value":{"type":"string"},"type":{"type":"string","enum":["IP","DOMAIN","URL","FILE_HASH"]},"threatSource":{"type":"string"},"confidence":{"type":"integer","format":"int32"},"malwareFamily":{"type":"string"},"campaign":{"type":"string"},"description":{"type":"string"}}},"IndicatorOfCompromise":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"value":{"type":"string"},"type":{"type":"string","enum":["IP","DOMAIN","URL","FILE_HASH"]},"threatSource":{"type":"string"},"confidence":{"type":"integer","format":"int32"},"malwareFamily":{"type":"string"},"campaign":{"type":"string"},"description":{"type":"string"},"active":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"Incident":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"timestamp":{"type":"string","format":"date-time"},"incidentType":{"type":"string"},"severity":{"type":"string"},"riskScore":{"type":"integer","format":"int32"},"summary":{"type":"string"},"sourceIp":{"type":"string"},"status":{"type":"string"},"assignedAnalyst":{"type":"string"},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"Alert":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"timestamp":{"type":"string","format":"date-time"},"alertType":{"type":"string"},"severity":{"type":"string"},"riskScore":{"type":"integer","format":"int32"},"message":{"type":"string"},"sourceIp":{"type":"string"},"username":{"type":"string"},"eventType":{"type":"string"},"mitreTechnique":{"type":"string"},"mitreDescription":{"type":"string"},"status":{"type":"string"},"assignedAnalyst":{"type":"string"},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"LogRequest":{"type":"object","properties":{"logLine":{"type":"string"}}},"SecurityEvent":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"timestamp":{"type":"string","format":"date-time"},"sourceIp":{"type":"string"},"username":{"type":"string"},"eventType":{"type":"string"},"rawLog":{"type":"string"},"destinationPort":{"type":"integer","format":"int32"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"username":{"type":"string"},"role":{"type":"string"}}},"PageSecurityEvent":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/SecurityEvent"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"DashboardSummary":{"type":"object","properties":{"totalEvents":{"type":"integer","format":"int64"},"totalAlerts":{"type":"integer","format":"int64"},"criticalAlerts":{"type":"integer","format":"int64"},"highAlerts":{"type":"integer","format":"int64"},"failedLogins":{"type":"integer","format":"int64"},"portScans":{"type":"integer","format":"int64"},"privilegeEscalations":{"type":"integer","format":"int64"},"totalRiskScore":{"type":"integer","format":"int32"},"averageRiskScore":{"type":"number","format":"double"},"highestRiskScore":{"type":"integer","format":"int32"},"criticalRiskAlertCount":{"type":"integer","format":"int64"},"highRiskAlertCount":{"type":"integer","format":"int64"}}},"DashboardTrends":{"type":"object","properties":{"alertsLast24Hours":{"type":"integer","format":"int64"},"alertsLast7Days":{"type":"integer","format":"int64"},"alertsLast30Days":{"type":"integer","format":"int64"},"eventsLast24Hours":{"type":"integer","format":"int64"},"eventsLast7Days":{"type":"integer","format":"int64"},"eventsLast30Days":{"type":"integer","format":"int64"}}},"TopUsername":{"type":"object","properties":{"username":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"TopSourceIp":{"type":"object","properties":{"sourceIp":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"TopAlertType":{"type":"object","properties":{"alertType":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"DashboardTimelinePoint":{"type":"object","properties":{"date":{"type":"string","format":"date"},"events":{"type":"integer","format":"int64"},"alerts":{"type":"integer","format":"int64"}}},"RiskDistribution":{"type":"object","properties":{"low":{"type":"integer","format":"int64"},"medium":{"type":"integer","format":"int64"},"high":{"type":"integer","format":"int64"},"critical":{"type":"integer","format":"int64"}}},"MitreStatistic":{"type":"object","properties":{"technique":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"IocStatistics":{"type":"object","properties":{"totalIocAlerts":{"type":"integer","format":"int64"},"criticalIocAlerts":{"type":"integer","format":"int64"},"uniqueMaliciousIps":{"type":"integer","format":"int64"}}},"AuditLog":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"action":{"type":"string","enum":["USER_REGISTERED","USER_LOGIN_SUCCESS","ALERT_ASSIGNED","ALERT_STATUS_CHANGED","ALERT_NOTES_UPDATED"]},"username":{"type":"string"},"details":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"PageAuditLog":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AuditLog"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageAlert":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/Alert"}},"number":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}